requirements/requirements.txt
Django==3.2.4
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
django-debug-toolbar==3.7.0
python-decouple==3.6
django-extensions==3.2.1
pytest==5.4.3
pytest-django==3.8.0
pytest-cov==2.10.0
six==1.14.0
whitenoise==6.2.0
cryptography==40.0.1
Pillow==10.0.0
requests==2.31.0
